requirements.txt
# Core frameworks
chainlit>=1.3.0
fastapi>=0.100.0
uvicorn[standard]>=0.30.0

# LangChain
langchain==1.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-mcp-adapters>=0.1.0

# LlamaIndex for RAG
llama-index>=0.12.0
llama-index-core>=0.12.0
llama-index-vector-stores-weaviate>=1.4.0
llama-index-embeddings-openai>=0.3.0

# MCP Server
fastmcp>=2.14.0

# Vector store
weaviate-client>=4.9.0

# OpenAI
openai>=1.50.0

# Utilities
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.11.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
